% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_to_assistant.R
\name{add_file_to_assistant}
\alias{add_file_to_assistant}
\title{Add file to assistant
This will only add the file if it not exists}
\usage{
add_file_to_assistant(assistant, file_id)
}
\arguments{
\item{assistant}{The assistant to add the file to}

\item{file_id}{The file id to add to the assistant}
}
\value{
The assistant object
}
\description{
Add file to assistant
This will only add the file if it not exists
}
