% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_add_file.R
\name{assistant_add_file}
\alias{assistant_add_file}
\title{Add file to assistant
This will only add the file if it not exists}
\usage{
assistant_add_file(assistant, file_id)
}
\arguments{
\item{assistant}{The assistant to add the file to}

\item{file_id}{The file id to add to the assistant}
}
\value{
The updated assistant object
}
\description{
Add file to assistant
This will only add the file if it not exists
}
