% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_create.R
\name{assistant_create}
\alias{assistant_create}
\title{Create an assistant}
\usage{
assistant_create(
  name,
  instructions,
  model = "gpt-4-turbo-preview",
  tool_retrieval = FALSE,
  tool_code_interpreter = FALSE,
  file_ids = NULL
)
}
\arguments{
\item{name}{The name of the assistant}

\item{instructions}{The instructions for the assistant}

\item{model}{The model to use for the assistant (default: "gpt-4-turbo-preview")}

\item{tool_retrieval}{Should the assistant have a retrieval tool? (default: FALSE)}

\item{tool_code_interpreter}{Should the assistant have a code interpreter tool? (default: FALSE)}

\item{file_ids}{A list of file IDs to add to the assistant (default: NULL)}
}
\value{
an assistant object
}
\description{
Create an assistant
}
