% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_delete.R
\name{assistant_delete}
\alias{assistant_delete}
\title{Delete an assistant by id}
\usage{
assistant_delete(assistant_id)
}
\arguments{
\item{assistant_id}{character id of the assistant}
}
\value{
A respsonse object with a field deleted set to TRUE if the assistant was deleted
}
\description{
Delete an assistant by id
}
\examples{
\dontrun{
assistant_delete("asst_abc123")
}
}
