% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file_create.R
\name{assistant_file_create}
\alias{assistant_file_create}
\title{Create Assistant file from assistant object and file id
file must already exist}
\usage{
assistant_file_create(assistant, file_id)
}
\arguments{
\item{assistant}{The assistant to associate the file with}

\item{file_id}{The file id to associate with the assistant}
}
\value{
An assistant_file object
}
\description{
Create Assistant file from assistant object and file id
file must already exist
}
