% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_file_delete.R
\name{assistant_file_delete}
\alias{assistant_file_delete}
\title{Delete an assistant file}
\usage{
assistant_file_delete(assistant, file_id)
}
\arguments{
\item{assistant}{The assistant to delete the file from}

\item{file_id}{The file id to delete}
}
\value{
The assistant object
}
\description{
Delete an assistant file
}
\examples{
\dontrun{
assistant_file_delete(assistant, file_id)
}
}
