% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_list.R
\name{assistant_list}
\alias{assistant_list}
\title{Get list of all assistants}
\usage{
assistant_list(order = "desc", limit = 20)
}
\arguments{
\item{order}{The order to list the assistants in. Options are "asc" and "desc".}

\item{limit}{The maximum number of assistants to return.}
}
\value{
A list of assistant objects
}
\description{
Get list of all assistants
}
\examples{
\dontrun{
assistant_list()
}
}
