% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_retrieve.R
\name{assistant_retrieve}
\alias{assistant_retrieve}
\title{Retrieves an assistant by id}
\usage{
assistant_retrieve(assistant_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant to retrieve.}
}
\value{
An assistant object
}
\description{
Retrieves an assistant by id
}
\examples{
\dontrun{
retrieveAssistant("asst_abc123")
}
}
