% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_run.R
\name{assistant_run}
\alias{assistant_run}
\title{Function that runs an assistant with a thread}
\usage{
assistant_run(assistant, thread)
}
\arguments{
\item{assistant}{The assistant to run}

\item{thread}{The thread to run the assistant with}
}
\value{
A run object
}
\description{
Function that runs an assistant with a thread
}
\examples{
\dontrun{
assistant <- create_assistant()
thread <- create_thread()
run <- run_assistant(assistant, thread)
if(run$status != "completed"){
 print("The assistant is still running")
}
}
}
