% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_token.R
\name{check_token}
\alias{check_token}
\title{Check whether the OpenAI Token is set}
\usage{
check_token()
}
\value{
TRUE is the token is set, errors otherwise
}
\description{
Check whether the OpenAI Token is set
}
\examples{
\dontrun{
check_token()
}
}
