% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assistant_file.R
\name{create_assistant_file}
\alias{create_assistant_file}
\title{Create Assistant file from assistant object and file id
curl https://api.openai.com/v1/assistants/asst_abc123/files \
-H 'Authorization: Bearer $OPENAI_API_KEY"' \
-H 'Content-Type: application/json' \
-H 'OpenAI-Beta: assistants=v1' \
-d '{
"file_id": "file-abc123"
}'}
\usage{
create_assistant_file(assistant, file_id)
}
\arguments{
\item{assistant}{The assistant to associate the file with}

\item{file_id}{The file id to associate with the assistant}
}
\value{
A file object
}
\description{
Create Assistant file from assistant object and file id
curl https://api.openai.com/v1/assistants/asst_abc123/files \
-H 'Authorization: Bearer $OPENAI_API_KEY"' \
-H 'Content-Type: application/json' \
-H 'OpenAI-Beta: assistants=v1' \
-d '{
"file_id": "file-abc123"
}'
}
