% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_thread.R
\name{create_thread}
\alias{create_thread}
\title{Create threadBeta
POST
https://api.openai.com/v1/threads
Create a thread.
curl https://api.openai.com/v1/threads \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1" \
-d ''}
\usage{
create_thread()
}
\value{
A thread object
}
\description{
Create threadBeta
POST
https://api.openai.com/v1/threads
Create a thread.
curl https://api.openai.com/v1/threads \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1" \
-d ''
}
\examples{
\dontrun{
thread <- create_thread()
}
}
