% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file.R
\name{deleteFile}
\alias{deleteFile}
\title{Delete a file by id
curl https://api.openai.com/v1/files/file-abc123 \
-X DELETE \
-H "Authorization: Bearer $OPENAI_API_KEY"}
\usage{
deleteFile(file_id)
}
\arguments{
\item{file_id}{The file id to delete}
}
\value{
A list of file objects?
}
\description{
Delete a file by id
curl https://api.openai.com/v1/files/file-abc123 \
-X DELETE \
-H "Authorization: Bearer $OPENAI_API_KEY"
}
