% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_delete.R
\name{file_delete}
\alias{file_delete}
\title{Delete a file by id}
\usage{
file_delete(file_id)
}
\arguments{
\item{file_id}{The file id to delete}
}
\value{
A response object with a field deleted set to TRUE if the file was deleted
}
\description{
Delete a file by id
}
\examples{
\dontrun{
file_delete("file_abc123")
}
}
