% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_message.R
\name{get_last_message}
\alias{get_last_message}
\title{Gets the content of the last message from a list of messages}
\usage{
get_last_message(messages)
}
\arguments{
\item{messages}{A list of message objects}
}
\value{
The content of the last message
}
\description{
Gets the content of the last message from a list of messages
}
