% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_active.R
\name{is_active}
\alias{is_active}
\title{Checks whether a run is active}
\usage{
is_active(run)
}
\arguments{
\item{run}{The run object to check}
}
\value{
TRUE if the run is active, FALSE otherwise
}
\description{
Checks whether a run is active
}
