% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_assistant.R
\name{is_assistant}
\alias{is_assistant}
\title{determines if a object is an assistant}
\usage{
is_assistant(assistant)
}
\arguments{
\item{assistant}{The object to check}
}
\value{
TRUE when the object is an assistant, FALSE otherwise
}
\description{
determines if a object is an assistant
}
