% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_assistant.R
\name{is_assistant_id}
\alias{is_assistant_id}
\title{verifies an assistant id as a valid assistant id}
\usage{
is_assistant_id(assistant_id)
}
\arguments{
\item{assistant_id}{The assistant id to verify}
}
\value{
TRUE if the assistant id is valid, FALSE otherwise
}
\description{
verifies an assistant id as a valid assistant id
}
