% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_file.R
\name{is_file_id}
\alias{is_file_id}
\title{Check if a string is a file id}
\usage{
is_file_id(file_id)
}
\arguments{
\item{file_id}{The file id to check}
}
\value{
TRUE if the file id is a file id, FALSE otherwise
}
\description{
Check if a string is a file id
}
