% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_run.R
\name{is_run}
\alias{is_run}
\title{Check if a run is a run}
\usage{
is_run(run)
}
\arguments{
\item{run}{The run object to check}
}
\value{
TRUE if the run is a run, FALSE otherwise
}
\description{
Check if a run is a run
}
