% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_run.R
\name{is_run_id}
\alias{is_run_id}
\title{verifies a run id}
\usage{
is_run_id(run_id)
}
\arguments{
\item{run_id}{The run id to verify}
}
\value{
TRUE if the run id is valid, FALSE otherwise
}
\description{
verifies a run id
}
