% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_thread.R
\name{is_thread}
\alias{is_thread}
\title{determines if object is a thread object}
\usage{
is_thread(thread)
}
\arguments{
\item{thread}{The object to check}
}
\value{
TRUE if the object is a thread object, FALSE otherwise
}
\description{
determines if object is a thread object
}
