% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_thread.R
\name{is_thread_id}
\alias{is_thread_id}
\title{verifies a thread id as a valid thread id}
\usage{
is_thread_id(thread_id)
}
\arguments{
\item{thread_id}{The thread id to verify}
}
\value{
TRUE if the thread id is valid, FALSE otherwise
}
\description{
verifies a thread id as a valid thread id
}
