% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assistants.R
\name{list_assistants}
\alias{list_assistants}
\title{Get list of all assistants
curl "https://api.openai.com/v1/assistants?order=desc&limit=20" \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1"}
\usage{
list_assistants(order = "desc", limit = 20)
}
\arguments{
\item{order}{The order to list the assistants in. Options are "asc" and "desc".}

\item{limit}{The maximum number of assistants to return.}
}
\value{
A list of assistant objects
}
\description{
Get list of all assistants
curl "https://api.openai.com/v1/assistants?order=desc&limit=20" \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1"
}
\examples{
\dontrun{
list_assistants()
}
}
