% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List all files
curl https://api.openai.com/v1/files \
-H "Authorization: Bearer $OPENAI_API_KEY"}
\usage{
list_files()
}
\value{
A list of file objects
}
\description{
List all files
curl https://api.openai.com/v1/files \
-H "Authorization: Bearer $OPENAI_API_KEY"
}
