% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_messages.R
\name{retrieve_messages}
\alias{retrieve_messages}
\title{Retrieve messages from thread object
curl https://api.openai.com/v1/threads/thread_abc123/messages \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1"}
\usage{
retrieve_messages(thread)
}
\arguments{
\item{thread}{The thread object to retrieve the messages from}
}
\value{
A list of message objects
}
\description{
Retrieve messages from thread object
curl https://api.openai.com/v1/threads/thread_abc123/messages \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1"
}
