% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_assistants.R
\name{retrieveAssistant}
\alias{retrieveAssistant}
\title{Retrieve assistant using the API
GET
https://api.openai.com/v1/assistants/{assistant_id}
Retrieves an assistant.
Path parameters
assistant_id
string
Required
The ID of the assistant to retrieve.}
\usage{
retrieveAssistant(assistant_id)
}
\arguments{
\item{assistant_id}{The ID of the assistant to retrieve.}
}
\value{
An assistant object
}
\description{
curl https://api.openai.com/v1/assistants/asst_abc123 \
-H "Content-Type: application/json" \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "OpenAI-Beta: assistants=v1"
}
\examples{
retrieveAssistant("asst_abc123")
}
