% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assistant.R
\name{run_assistant}
\alias{run_assistant}
\title{function that runs an assistant with a thread
curl https://api.openai.com/v1/threads/thread_abc123/runs \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "Content-Type: application/json" \
-H "OpenAI-Beta: assistants=v1" \
-d '{
"assistant_id": "asst_abc123"
}'}
\usage{
run_assistant(thread, assistant)
}
\arguments{
\item{thread}{The thread to run the assistant with}

\item{assistant}{The assistant to run}
}
\value{
A run object
}
\description{
function that runs an assistant with a thread
curl https://api.openai.com/v1/threads/thread_abc123/runs \
-H "Authorization: Bearer $OPENAI_API_KEY" \
-H "Content-Type: application/json" \
-H "OpenAI-Beta: assistants=v1" \
-d '{
"assistant_id": "asst_abc123"
}'
}
\examples{
\dontrun{
thread <- create_thread()
assistant <- create_assistant()
run <- run_assistant(thread, assistant)
if(run$status != "completed"){
 print("The assistant is still running")
}
}
}
