% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_retrieve.R
\name{run_retrieve}
\alias{run_retrieve}
\title{Retrieve a run by thread and run object}
\usage{
run_retrieve(run, thread)
}
\arguments{
\item{run}{The run object to retrieve}

\item{thread}{The thread object to retrieve the run from}
}
\value{
The updated run object
}
\description{
Retrieve a run by thread and run object
}
\examples{
\dontrun{
# more details here TBD
run_retrieve(run, thread)
}
}
