% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_retrieve_by_id.R
\name{run_retrieve_by_id}
\alias{run_retrieve_by_id}
\title{Retrieve a run by thread and run object by ID}
\usage{
run_retrieve_by_id(run_id, thread_id)
}
\arguments{
\item{run_id}{The run id to retrieve}

\item{thread_id}{The thread id to retrieve the run from}
}
\value{
The updated run object
}
\description{
Retrieve a run by thread and run object by ID
}
\examples{
\dontrun{
run_retrieve_by_id("run_abc123", "thread_abc123")
}
}
