% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_add_message.R
\name{thread_add_message}
\alias{thread_add_message}
\title{Add a message to a thread}
\usage{
thread_add_message(thread, message)
}
\arguments{
\item{thread}{A thread object}

\item{message}{The text of the message to add as a string}
}
\value{
A message object
}
\description{
Add a message to a thread
}
\examples{
\dontrun{
thread <- create_thread()
msg <- thread_add_message(thread, "How does AI work? Explain it in simple terms.")
}
}
