% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_create.R
\name{thread_create}
\alias{thread_create}
\title{Create a new thread}
\usage{
thread_create()
}
\value{
A thread object
}
\description{
Create a new thread
}
\examples{
\dontrun{
thread <- thread_create()
}
}
