% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_retrieve.R
\name{thread_retrieve}
\alias{thread_retrieve}
\title{Retrieve thread by id}
\usage{
thread_retrieve(thread_id)
}
\arguments{
\item{thread_id}{The ID of the thread to retrieve}
}
\value{
The thread object
}
\description{
Retrieve thread by id
}
