% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_retrieve_messages.R
\name{thread_retrieve_messages}
\alias{thread_retrieve_messages}
\title{Retrieve messages from thread object}
\usage{
thread_retrieve_messages(thread)
}
\arguments{
\item{thread}{The thread object to retrieve the messages from}
}
\value{
A list of message objects
}
\description{
Retrieve messages from thread object
}
\examples{
\dontrun{
thread <- create_thread()
messages <- thread_retrieve_messages(thread)
}
}
